backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
boto3==1.26.79
botocore==1.29.79
Django==4.1.4
django-cors-headers==3.13.0
django-extensions==3.2.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.4.0
pycodestyle==2.10.0
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
